Backend/requirements.txt
jiwer==2.5.1
keras==2.10.0
numpy==1.21.5
opencv_python==4.6.0.66
path.py==12.5.0
tensorflow==2.10.0
